+++ common/src/main/java/com/sonicether/soundphysics/mixin/PlaySoundCommandMixin.java
@@ -1,15 +1,13 @@
 package com.sonicether.soundphysics.mixin;
 
-import com.sonicether.soundphysics.SoundPhysicsMod;
 import net.minecraft.server.commands.PlaySoundCommand;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.Constant;
-import org.spongepowered.asm.mixin.injection.ModifyConstant;
 
 @Mixin(PlaySoundCommand.class)
 public class PlaySoundCommandMixin {
 
-    @ModifyConstant(method = "playSound", constant = @Constant(floatValue = 16F), expect = 1)
+    //TODO Fix
+    /*@ModifyConstant(method = "playSound", constant = @Constant(floatValue = 16F), expect = 1)
     private static float allowance1(float value) {
         if (!SoundPhysicsMod.CONFIG.enabled.get()) {
             return value;
@@ -23,6 +21,6 @@ public class PlaySoundCommandMixin {
             return value;
         }
         return value * SoundPhysicsMod.CONFIG.soundDistanceAllowance.get();
-    }
+    }*/
 
 }