+++ forge/changelog.md
@@ -1,1 +1,1 @@
-- Updated to 1.19.2
+- Fixed errors when having environment_evaluation_ray_bounces set to a lower value than 4